Webpack support - don't use require.resolve(), allow user to use it
The README and in-code comment instructs users they may need to use
require.resolve() to provide the correct path to requireTaskPool.

This is a problem for Webpack users because Webpack rewrites
require.resolve().

Removing require.resolve() enables Webpack users to use
electron-remote.

I'm not sure, but I think for some users who implicitly depend on the
current behavior of electron-remote require.resolve()-ing for them,
this may be a breaking change.